Count Issue, PR, PR review, and Commit contributions
BREAKING CHANGE:
The old API, giving extensive information about PRs is now exported as
`pullRequests`. The added API, giving summary information about all
types of contributions, is exported as `contributions`
