+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+## v2.0.0
+
+- Add a way to get stats about various types of contributions, not just PRs.
+  BREAKING CHANGE: The old API, giving extensive information about PRs is now exported as `pullRequests`.
+  The added API, giving summary information about all types of contributions, is exported as `contributions`.
