+++ package.json
@@ -11,8 +11,11 @@
   "author": "YLD Open Source (opensource-maintainers@yld.io)",
   "license": "Apache-2.0",
   "dependencies": {
+    "date-fns": "^2.12.0",
     "github-graphql-api": "^1.0.6",
-    "lodash": "^4.17.15"
+    "got": "^11.0.3",
+    "lodash": "^4.17.15",
+    "p-retry": "^4.2.0"
   },
   "devDependencies": {
     "express": "^4.17.1",
